+++ src/gdmbayes/plotting/__init__.py
@@ -1,19 +1,11 @@
 """Plotting utilities for spGDMM models."""
 
-from .plots import (
-    plot_crps_comparison,
-    plot_isplines,
-    plot_ppc,
-    rgb_biological_space,
-    rgb_from_biological_space,
-    summarise_sampling,
-)
+from .isplines import plot_isplines
+from .ppc import plot_ppc
+from .scoring import crps_boxplot
 
 __all__ = [
     "plot_isplines",
-    "plot_crps_comparison",
-    "summarise_sampling",
     "plot_ppc",
-    "rgb_from_biological_space",
-    "rgb_biological_space",
+    "crps_boxplot",
 ]